refactor: update component imports and optimize carousel state
Replace barrel imports with direct component imports to improve tree-shaking and maintainability. Remove unused state variables in ProjectShowcase to simplify carousel logic. Add passive flag to scroll event listener for better performance. Add unique ID to JSON-LD script to prevent duplicate execution.
