+++ apps/gmozer.ca/src/app/[locale]/(footer)/contact/page.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import { getTranslations } from "next-intl/server";
-import { ContactForm } from "@/components";
+import { ContactForm } from "@/components/contact-form";
 
 export async function generateMetadata() {
 	const t = await getTranslations("metadata");
