+++ apps/gmozer.ca/src/app/[locale]/(footer)/work/page.tsx
@@ -1,5 +1,5 @@
 import { getTranslations } from "next-intl/server";
-import { ProjectShowcase } from "@/components";
+import { ProjectShowcase } from "@/components/project-showcase";
 
 export async function generateMetadata() {
 	const t = await getTranslations("metadata");
