+++ apps/gmozer.ca/src/components/header/header.tsx
@@ -25,7 +25,7 @@ export const Header = () => {
 			setIsTransparent(window.scrollY === 0);
 		};
 
-		document.addEventListener("scroll", handleScroll);
+		document.addEventListener("scroll", handleScroll, { passive: true });
 
 		return () => document.removeEventListener("scroll", handleScroll);
 	}, []);
